# swagger.yaml
openapi: 3.0.3
info:
  title: PowerChain AI Sovereign Grid Telemetry & Treasury API
  description: OpenAPI 3.0 specification for PowerChain AI dePIN grid intelligence, Pyth oracles, Solana Pay, and PWRC token credit clearing.
  version: 1.2.0
paths:
...
  /api/v1/telemetry/nodes:
    get:
      summary: Detailed Node Hardware & Sensor Diagnostic
      parameters:
        - name: id
          in: query
          description: Target dePIN node identifier
          schema:
            type: string
      responses:
        '200':
          description: Single node voltage, current, frequency, and BESS state-of-charge metrics
  /api/v1/search:
    get:
      summary: Global Sovereign DePIN & LLM Search
      parameters:
        - name: q
          in: query
          description: Search term for nodes, prompts, assets, or models
          schema:
            type: string
      responses:
        '200':
          description: Instant search results matching telemetry, workflows, assets, and AI providers
...